feat: Implementations of Utility for testing the memory subsystem
This is a userspace utility for testing the memory subsystem for faults. In comparison to memtest86 you do not need to reboot the computer to test for memory faults.

Memtester can also be told to test memory starting at a particular physical address.

Log: Implementations of Utility for testing the memory subsystem

Issue: https://github.com/deepin-community/sig-deepin-sysdev-team/issues/297